Completed user stories 1 and 2

+++ src/main/java/com/model/Employee.java
@@ -16,7 +16,7 @@ public class Employee {
 	
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long empId;
+    private int empId;
     private String name;
     private String username;
     private String password;
